Hide the realname field from the preferences form
The recently added solution based on a callback
AuthMoodle::allowRealNameChange() does not seem to work that well. It
led to a situation that the user.user_real_name in the database was
emptied every time the user saved the preferences form.

This has a drawback that the real name is not longer displayed in the
preferences form (as a static text) but it does not matter at all imho).