Improve the navigation bar display
Some of our wiki provides custom navigation trail (typically via the template).
Instead of having such wikis hardcoded in the code, we now check is such a
custom navtrail is present. If not, the default navigation bar is displayed.

The default navigation bar has been simplified to look more like the custom one
we use.