Updated mediawiki/core Project: mediawiki/extensions/CirrusSearch  5d2f813af4df1efc064524c2e86385db6b323b37
Suppress interwiki results when they would break

Sometimes we end up with a SearchResults object that can't handle interwiki
results and it just isn't fair to try to feed them to it. It makes fatals.

There is another fix for this issue, but its more complex and as of yet
doesn't pass all the tests. But this is a simpler, more obvious fix we can
can use as a stopgap.

Bug: T104189
Change-Id: Id46545dd77a00ccd4ba2f18efc5d7d57f2d96626
(cherry picked from commit 037592809fc60a9f39f92a60f611811f507e077b)